pyproject.toml
[project]
name = "hexlet-python-seo-uv"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.8",
    "validators>=0.35.0",
]
